for #660: refactor asserts
