+++ sharding-core/src/test/java/io/shardingjdbc/core/parsing/integrate/engine/IntegrateSupportedSQLParsingTest.java
@@ -17,66 +17,18 @@
 
 package io.shardingjdbc.core.parsing.integrate.engine;
 
-import com.google.common.base.Optional;
 import io.shardingjdbc.core.constant.DatabaseType;
 import io.shardingjdbc.core.parsing.SQLParsingEngine;
 import io.shardingjdbc.core.parsing.integrate.asserts.ParserResultSetLoader;
-import io.shardingjdbc.core.parsing.integrate.jaxb.condition.ExpectedCondition;
-import io.shardingjdbc.core.parsing.integrate.jaxb.condition.ExpectedValue;
-import io.shardingjdbc.core.parsing.integrate.jaxb.groupby.ExpectedGroupByColumn;
-import io.shardingjdbc.core.parsing.integrate.jaxb.item.ExpectedAggregationSelectItem;
-import io.shardingjdbc.core.parsing.integrate.jaxb.limit.ExpectedLimit;
-import io.shardingjdbc.core.parsing.integrate.jaxb.orderby.ExpectedOrderByColumn;
-import io.shardingjdbc.core.parsing.integrate.jaxb.root.ParserResult;
-import io.shardingjdbc.core.parsing.integrate.jaxb.table.ExpectedTable;
-import io.shardingjdbc.core.parsing.integrate.jaxb.token.ExpectedGeneratedKeyToken;
-import io.shardingjdbc.core.parsing.integrate.jaxb.token.ExpectedIndexToken;
-import io.shardingjdbc.core.parsing.integrate.jaxb.token.ExpectedItemsToken;
-import io.shardingjdbc.core.parsing.integrate.jaxb.token.ExpectedMultipleInsertValuesToken;
-import io.shardingjdbc.core.parsing.integrate.jaxb.token.ExpectedOffsetToken;
-import io.shardingjdbc.core.parsing.integrate.jaxb.token.ExpectedOrderByToken;
-import io.shardingjdbc.core.parsing.integrate.jaxb.token.ExpectedRowCountToken;
-import io.shardingjdbc.core.parsing.integrate.jaxb.token.ExpectedTokens;
-import io.shardingjdbc.core.parsing.integrate.jaxb.token.ExpectedTableToken;
-import io.shardingjdbc.core.parsing.parser.context.OrderItem;
-import io.shardingjdbc.core.parsing.parser.context.condition.Column;
-import io.shardingjdbc.core.parsing.parser.context.condition.Condition;
-import io.shardingjdbc.core.parsing.parser.context.condition.Conditions;
-import io.shardingjdbc.core.parsing.parser.context.limit.Limit;
-import io.shardingjdbc.core.parsing.parser.context.selectitem.AggregationSelectItem;
-import io.shardingjdbc.core.parsing.parser.context.selectitem.SelectItem;
-import io.shardingjdbc.core.parsing.parser.context.table.Table;
-import io.shardingjdbc.core.parsing.parser.context.table.Tables;
-import io.shardingjdbc.core.parsing.parser.sql.SQLStatement;
-import io.shardingjdbc.core.parsing.parser.sql.dql.select.SelectStatement;
-import io.shardingjdbc.core.parsing.parser.token.GeneratedKeyToken;
-import io.shardingjdbc.core.parsing.parser.token.IndexToken;
-import io.shardingjdbc.core.parsing.parser.token.ItemsToken;
-import io.shardingjdbc.core.parsing.parser.token.MultipleInsertValuesToken;
-import io.shardingjdbc.core.parsing.parser.token.OffsetToken;
-import io.shardingjdbc.core.parsing.parser.token.OrderByToken;
-import io.shardingjdbc.core.parsing.parser.token.RowCountToken;
-import io.shardingjdbc.core.parsing.parser.token.SQLToken;
-import io.shardingjdbc.core.parsing.parser.token.TableToken;
+import io.shardingjdbc.core.parsing.integrate.asserts.SQLStatementAssert;
 import io.shardingjdbc.test.sql.SQLCaseType;
 import io.shardingjdbc.test.sql.SQLCasesLoader;
 import lombok.RequiredArgsConstructor;
 import org.junit.Test;
 import org.junit.runners.Parameterized.Parameters;
 
-import java.lang.reflect.Field;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertThat;
-import static org.junit.Assert.assertTrue;
 
 @RequiredArgsConstructor
 public final class IntegrateSupportedSQLParsingTest extends AbstractBaseIntegrateSQLParsingTest {
@@ -97,405 +49,8 @@ public final class IntegrateSupportedSQLParsingTest extends AbstractBaseIntegrat
     }
     
     @Test
-    public void assertSupportedSQL() throws NoSuchFieldException, IllegalAccessException {
+    public void assertSupportedSQL() {
         String sql = sqlCasesLoader.getSupportedSQL(sqlCaseId, sqlCaseType, parserResultSetLoader.getParserResult(sqlCaseId).getParameters());
-        assertSQLStatement(new SQLParsingEngine(databaseType, sql, getShardingRule()).parse());
-    }
-    
-    private void assertSQLStatement(final SQLStatement actual) throws NoSuchFieldException, IllegalAccessException {
-        ParserResult parserResult = parserResultSetLoader.getParserResult(sqlCaseId);
-        assertTables(actual.getTables(), parserResult.getTables());
-        assertConditions(actual.getConditions(), parserResult.getConditions());
-        assertSQLTokens(actual.getSqlTokens(), parserResult.getTokens());
-        if (actual instanceof SelectStatement) {
-            SelectStatement selectStatement = (SelectStatement) actual;
-            assertItems(selectStatement.getItems(), parserResult.getAggregationSelectItems());
-            assertGroupByItems(selectStatement.getGroupByItems(), parserResult.getGroupByColumns());
-            assertOrderByItems(selectStatement.getOrderByItems(), parserResult.getOrderByColumns());
-            assertLimit(selectStatement.getLimit(), parserResult.getLimit());
-        }
-    }
-    
-    private void assertTables(final Tables actual, final List<ExpectedTable> expected) {
-        assertThat(getFullAssertMessage("Tables size assertion error: "), actual.getTableNames().size(), is(expected.size()));
-        for (ExpectedTable each : expected) {
-            Optional<Table> table;
-            if (null != each.getAlias()) {
-                table = actual.find(each.getAlias());
-            } else {
-                table = actual.find(each.getName());
-            }
-            assertTrue(getFullAssertMessage("Table should exist: "), table.isPresent());
-            assertTable(table.get(), each);
-        }
-    }
-    
-    private void assertTable(final Table actual, final ExpectedTable expected) {
-        assertThat(getFullAssertMessage("Table name assertion error: "), actual.getName(), is(expected.getName()));
-        assertThat(getFullAssertMessage("Table alias assertion error: "), actual.getAlias().orNull(), is(expected.getAlias()));
-    }
-    
-    private void assertConditions(final Conditions actual, final List<ExpectedCondition> expected) throws NoSuchFieldException, IllegalAccessException {
-        assertThat(getFullAssertMessage("Conditions size assertion error: "), actual.getConditions().size(), is(expected.size()));
-        for (ExpectedCondition each : expected) {
-            Optional<Condition> condition = actual.find(new Column(each.getColumnName(), each.getTableName()));
-            assertTrue(getFullAssertMessage("Table should exist: "), condition.isPresent());
-            assertCondition(condition.get(), each);
-        }
-    }
-    
-    @SuppressWarnings("unchecked")
-    private void assertCondition(final Condition actual, final ExpectedCondition expected) throws NoSuchFieldException, IllegalAccessException {
-        assertThat(getFullAssertMessage("Condition column name assertion error: "), actual.getColumn().getName().toUpperCase(), is(expected.getColumnName().toUpperCase()));
-        assertThat(getFullAssertMessage("Condition table name assertion error: "), actual.getColumn().getTableName().toUpperCase(), is(expected.getTableName().toUpperCase()));
-        assertThat(getFullAssertMessage("Condition operator assertion error: "), actual.getOperator().name(), is(expected.getOperator()));
-        int count = 0;
-        for (ExpectedValue each : expected.getValues()) {
-            Map<Integer, Comparable<?>> positionValueMap = (Map<Integer, Comparable<?>>) getField(actual, "positionValueMap");
-            Map<Integer, Integer> positionIndexMap = (Map<Integer, Integer>) getField(actual, "positionIndexMap");
-            if (!positionValueMap.isEmpty()) {
-                assertThat(getFullAssertMessage("Condition parameter value assertion error: "), positionValueMap.get(count), is((Comparable) each.getLiteralForAccurateType()));
-            } else if (!positionIndexMap.isEmpty()) {
-                assertThat(getFullAssertMessage("Condition parameter index assertion error: "), positionIndexMap.get(count), is(each.getIndex()));
-            }
-            count++;
-        }
-    }
-    
-    private Object getField(final Object actual, final String fieldName) throws NoSuchFieldException, IllegalAccessException {
-        Field field = actual.getClass().getDeclaredField(fieldName);
-        field.setAccessible(true);
-        return field.get(actual);
-    }
-    
-    private void assertSQLTokens(final List<SQLToken> actual, final ExpectedTokens expected) {
-        assertTableTokens(actual, expected);
-        assertIndexToken(actual, expected);
-        assertItemsToken(actual, expected);
-        assertGeneratedKeyToken(actual, expected);
-        assertMultipleInsertValuesToken(actual, expected);
-        assertOrderByToken(actual, expected);
-        assertOffsetToken(actual, expected);
-        assertRowCountToken(actual, expected);
-    }
-    
-    private void assertTableTokens(final List<SQLToken> actual, final ExpectedTokens expected) {
-        List<TableToken> tableTokens = getTableTokens(actual);
-        assertThat(getFullAssertMessage("Table tokens size error: "), tableTokens.size(), is(expected.getTableTokens().size()));
-        int count = 0;
-        for (ExpectedTableToken each : expected.getTableTokens()) {
-            assertTableToken(tableTokens.get(count), each);
-            count++;
-        }
-    }
-    
-    private void assertTableToken(final TableToken actual, final ExpectedTableToken expected) {
-        assertThat(getFullAssertMessage("Table tokens begin position assertion error: "), actual.getBeginPosition(), is(expected.getBeginPosition()));
-        assertThat(getFullAssertMessage("Table tokens original literals assertion error: "), actual.getOriginalLiterals(), is(expected.getOriginalLiterals()));
-    }
-    
-    private List<TableToken> getTableTokens(final List<SQLToken> actual) {
-        List<TableToken> result = new ArrayList<>(actual.size());
-        for (SQLToken each : actual) {
-            if (each instanceof TableToken) {
-                result.add((TableToken) each);
-            }
-        }
-        return result;
-    }
-    
-    private void assertIndexToken(final List<SQLToken> actual, final ExpectedTokens expected) {
-        Optional<IndexToken> indexToken = getIndexToken(actual);
-        if (indexToken.isPresent()) {
-            assertIndexToken(indexToken.get(), expected.getIndexToken());
-        } else {
-            assertNull(getFullAssertMessage("Index token should not exist: "), expected.getIndexToken());
-        }
-    }
-    
-    private void assertIndexToken(final IndexToken actual, final ExpectedIndexToken expected) {
-        assertThat(getFullAssertMessage("Index token begin position assertion error: "), actual.getBeginPosition(), is(expected.getBeginPosition()));
-        assertThat(getFullAssertMessage("Index token original literals assertion error: "), actual.getOriginalLiterals(), is(expected.getOriginalLiterals()));
-        assertThat(getFullAssertMessage("Index token table name assertion error: "), actual.getTableName(), is(expected.getTableName()));
-    }
-    
-    private Optional<IndexToken> getIndexToken(final List<SQLToken> actual) {
-        for (SQLToken each : actual) {
-            if (each instanceof IndexToken) {
-                return Optional.of((IndexToken) each);
-            }
-        }
-        return Optional.absent();
-    }
-    
-    private void assertItemsToken(final List<SQLToken> actual, final ExpectedTokens expected) {
-        Optional<ItemsToken> itemsToken = getItemsToken(actual);
-        if (itemsToken.isPresent()) {
-            assertItemsToken(itemsToken.get(), expected.getItemsToken());
-        } else {
-            assertNull(getFullAssertMessage("Items token should not exist: "), expected.getItemsToken());
-        }
-    }
-    
-    private void assertItemsToken(final ItemsToken actual, final ExpectedItemsToken expected) {
-        assertThat(getFullAssertMessage("Items token begin position assertion error: "), actual.getBeginPosition(), is(expected.getBeginPosition()));
-        assertThat(getFullAssertMessage("Items token items assertion error: "), actual.getItems(), is(expected.getItems()));
-    }
-    
-    private Optional<ItemsToken> getItemsToken(final List<SQLToken> actual) {
-        for (SQLToken each : actual) {
-            if (each instanceof ItemsToken) {
-                return Optional.of((ItemsToken) each);
-            }
-        }
-        return Optional.absent();
-    }
-    
-    private void assertGeneratedKeyToken(final List<SQLToken> actual, final ExpectedTokens expected) {
-        Optional<GeneratedKeyToken> generatedKeyToken = getGeneratedKeyToken(actual);
-        if (generatedKeyToken.isPresent()) {
-            assertGeneratedKeyToken(generatedKeyToken.get(), expected.getGeneratedKeyToken());
-        } else {
-            assertNull(getFullAssertMessage("Generated key token should not exist: "), expected.getGeneratedKeyToken());
-        }
-    }
-    
-    private void assertGeneratedKeyToken(final GeneratedKeyToken actual, final ExpectedGeneratedKeyToken expected) {
-        if (SQLCaseType.Placeholder == sqlCaseType) {
-            assertThat(getFullAssertMessage("Generated key token begin position assertion error: "), actual.getBeginPosition(), is(expected.getPlaceholderBeginPosition()));
-        } else {
-            assertThat(getFullAssertMessage("Generated key token begin position assertion error: "), actual.getBeginPosition(), is(expected.getLiteralBeginPosition()));
-        }
-    }
-    
-    private Optional<GeneratedKeyToken> getGeneratedKeyToken(final List<SQLToken> actual) {
-        for (SQLToken each : actual) {
-            if (each instanceof GeneratedKeyToken) {
-                return Optional.of((GeneratedKeyToken) each);
-            }
-        }
-        return Optional.absent();
-    }
-    
-    private void assertMultipleInsertValuesToken(final List<SQLToken> actual, final ExpectedTokens expected) {
-        Optional<MultipleInsertValuesToken> multipleInsertValuesToken = getMultipleInsertValuesToken(actual);
-        if (multipleInsertValuesToken.isPresent()) {
-            assertMultipleInsertValuesToken(multipleInsertValuesToken.get(), expected.getMultipleInsertValuesToken());
-        } else {
-            assertNull(getFullAssertMessage("Multiple insert values token should not exist: "), expected.getMultipleInsertValuesToken());
-        }
-    }
-    
-    private void assertMultipleInsertValuesToken(final MultipleInsertValuesToken actual, final ExpectedMultipleInsertValuesToken expected) {
-        assertThat(getFullAssertMessage("Multiple insert values token begin position assertion error: "), actual.getBeginPosition(), is(expected.getBeginPosition()));
-        assertThat(getFullAssertMessage("Multiple insert values token values assertion error: "), actual.getValues(), is(expected.getValues()));
-    }
-    
-    private Optional<MultipleInsertValuesToken> getMultipleInsertValuesToken(final List<SQLToken> actual) {
-        for (SQLToken each : actual) {
-            if (each instanceof MultipleInsertValuesToken) {
-                return Optional.of((MultipleInsertValuesToken) each);
-            }
-        }
-        return Optional.absent();
-    }
-    
-    private void assertOrderByToken(final List<SQLToken> actual, final ExpectedTokens expected) {
-        Optional<OrderByToken> orderByToken = getOrderByToken(actual);
-        if (orderByToken.isPresent()) {
-            assertOrderByToken(orderByToken.get(), expected.getOrderByToken());
-        } else {
-            assertNull(getFullAssertMessage("Order by token should not exist: "), expected.getOrderByToken());
-        }
-    }
-    
-    private void assertOrderByToken(final OrderByToken actual, final ExpectedOrderByToken expected) {
-        if (SQLCaseType.Placeholder == sqlCaseType) {
-            assertThat(getFullAssertMessage("Order by token begin position assertion error: "), actual.getBeginPosition(), is(expected.getPlaceholderBeginPosition()));
-        } else {
-            assertThat(getFullAssertMessage("Order by token begin position assertion error: "), actual.getBeginPosition(), is(expected.getLiteralBeginPosition()));
-        }
-    }
-    
-    private Optional<OrderByToken> getOrderByToken(final List<SQLToken> actual) {
-        for (SQLToken each : actual) {
-            if (each instanceof OrderByToken) {
-                return Optional.of((OrderByToken) each);
-            }
-        }
-        return Optional.absent();
-    }
-    
-    private void assertOffsetToken(final List<SQLToken> actual, final ExpectedTokens expected) {
-        Optional<OffsetToken> offsetToken = getOffsetToken(actual);
-        if (SQLCaseType.Placeholder == sqlCaseType) {
-            assertFalse(getFullAssertMessage("Offset token should not exist: "), offsetToken.isPresent());
-            return;
-        }
-        if (offsetToken.isPresent()) {
-            assertOffsetToken(offsetToken.get(), expected.getOffsetToken());
-        } else {
-            assertNull(getFullAssertMessage("Offset token should not exist: "), expected.getOffsetToken());
-        }
-    }
-    
-    private void assertOffsetToken(final OffsetToken actual, final ExpectedOffsetToken expected) {
-        assertThat(getFullAssertMessage("Offset token begin position assertion error: "), actual.getBeginPosition(), is(expected.getBeginPosition()));
-        assertThat(getFullAssertMessage("Offset token offset assertion error: "), actual.getOffset(), is(expected.getOffset()));
-    }
-    
-    private Optional<OffsetToken> getOffsetToken(final List<SQLToken> actual) {
-        for (SQLToken each : actual) {
-            if (each instanceof OffsetToken) {
-                return Optional.of((OffsetToken) each);
-            }
-        }
-        return Optional.absent();
-    }
-    
-    private void assertRowCountToken(final List<SQLToken> actual, final ExpectedTokens expected) {
-        Optional<RowCountToken> rowCountToken = getRowCountToken(actual);
-        if (SQLCaseType.Placeholder == sqlCaseType) {
-            assertFalse(getFullAssertMessage("Row count token should not exist: "), rowCountToken.isPresent());
-            return;
-        }
-        if (rowCountToken.isPresent()) {
-            assertRowCountToken(rowCountToken.get(), expected.getRowCountToken());
-        } else {
-            assertNull(getFullAssertMessage("Row count token should not exist: "), expected.getRowCountToken());
-        }
-    }
-    
-    private void assertRowCountToken(final RowCountToken actual, final ExpectedRowCountToken expected) {
-        assertThat(getFullAssertMessage("Row count token begin position assertion error: "), actual.getBeginPosition(), is(expected.getBeginPosition()));
-        assertThat(getFullAssertMessage("Row count token row count assertion error: "), actual.getRowCount(), is(expected.getRowCount()));
-    }
-    
-    private Optional<RowCountToken> getRowCountToken(final List<SQLToken> actual) {
-        for (SQLToken each : actual) {
-            if (each instanceof RowCountToken) {
-                return Optional.of((RowCountToken) each);
-            }
-        }
-        return Optional.absent();
-    }
-    
-    private void assertItems(final Set<SelectItem> actual, final List<ExpectedAggregationSelectItem> expected) {
-        // TODO assert SelectItems total size
-        // TODO assert StarSelectItem
-        // TODO assert CommonSelectItem
-        assertAggregationSelectItems(actual, expected);
-    }
-    
-    private void assertAggregationSelectItems(final Set<SelectItem> actual, final List<ExpectedAggregationSelectItem> expected) {
-        List<AggregationSelectItem> aggregationSelectItems = getAggregationSelectItems(actual);
-        assertThat(getFullAssertMessage("Table tokens size error: "), aggregationSelectItems.size(), is(expected.size()));
-        int count = 0;
-        for (AggregationSelectItem each : aggregationSelectItems) {
-            assertAggregationSelectItem(each, expected.get(count));
-            count++;
-        }
-    }
-    
-    private void assertAggregationSelectItem(final AggregationSelectItem actual, final ExpectedAggregationSelectItem expected) {
-        assertThat(getFullAssertMessage("Aggregation select item aggregation type assertion error: "), actual.getType().name(), is(expected.getType()));
-        assertThat(getFullAssertMessage("Aggregation select item inner expression assertion error: "), actual.getInnerExpression(), is(expected.getInnerExpression()));
-        assertThat(getFullAssertMessage("Aggregation select item alias assertion error: "), actual.getAlias().orNull(), is(expected.getAlias()));
-        assertThat(getFullAssertMessage("Aggregation select item index assertion error: "), actual.getIndex(), is(expected.getIndex()));
-        assertThat(getFullAssertMessage("Aggregation select item derived aggregation select items assertion error: "), 
-                actual.getDerivedAggregationSelectItems().size(), is(expected.getDerivedColumns().size()));
-        int count = 0;
-        for (AggregationSelectItem each : actual.getDerivedAggregationSelectItems()) {
-            assertAggregationSelectItem(each, expected.getDerivedColumns().get(count));
-            count++;
-        }
-    }
-    
-    private List<AggregationSelectItem> getAggregationSelectItems(final Set<SelectItem> actual) {
-        List<AggregationSelectItem> result = new ArrayList<>(actual.size());
-        for (SelectItem each : actual) {
-            if (each instanceof AggregationSelectItem) {
-                result.add((AggregationSelectItem) each);
-            }
-        }
-        return result;
-    }
-    
-    private void assertGroupByItems(final List<OrderItem> actual, final List<ExpectedGroupByColumn> expected) {
-        assertThat(getFullAssertMessage("Group by items size error: "), actual.size(), is(expected.size()));
-        int count = 0;
-        for (OrderItem each : actual) {
-            assertGroupByItem(each, expected.get(count));
-            count++;
-        }
-    }
-    
-    private void assertGroupByItem(final OrderItem actual, final ExpectedGroupByColumn expected) {
-        assertThat(getFullAssertMessage("Group by item owner assertion error: "), actual.getOwner().orNull(), is(expected.getOwner()));
-        assertThat(getFullAssertMessage("Group by item name assertion error: "), actual.getName().orNull(), is(expected.getName()));
-        assertThat(getFullAssertMessage("Group by item order direction assertion error: "), actual.getOrderDirection().name(), is(expected.getOrderDirection()));
-        // TODO assert nullOrderDirection
-        assertThat(getFullAssertMessage("Group by item alias assertion error: "), actual.getAlias().orNull(), is(expected.getAlias()));
-    }
-    
-    private void assertOrderByItems(final List<OrderItem> actual, final List<ExpectedOrderByColumn> expected) {
-        assertThat(getFullAssertMessage("Order by items size error: "), actual.size(), is(expected.size()));
-        int count = 0;
-        for (OrderItem each : actual) {
-            assertOrderByItem(each, expected.get(count));
-            count++;
-        }
-    }
-    
-    private void assertOrderByItem(final OrderItem actual, final ExpectedOrderByColumn expected) {
-        assertThat(getFullAssertMessage("Order by item owner assertion error: "), actual.getOwner().orNull(), is(expected.getOwner()));
-        assertThat(getFullAssertMessage("Order by item name assertion error: "), actual.getName().orNull(), is(expected.getName()));
-        assertThat(getFullAssertMessage("Order by item order direction assertion error: "), actual.getOrderDirection().name(), is(expected.getOrderDirection()));
-        // TODO assert nullOrderDirection
-        assertThat(getFullAssertMessage("Order by item index assertion error: "), actual.getIndex(), is(expected.getIndex()));
-        assertThat(getFullAssertMessage("Order by item alias assertion error: "), actual.getAlias().orNull(), is(expected.getAlias()));
-    }
-    
-    private void assertLimit(final Limit actual, final ExpectedLimit expected) {
-        if (null == actual) {
-            assertNull(getFullAssertMessage("Limit should not exist: "), expected);
-            return;
-        }
-        if (SQLCaseType.Placeholder == sqlCaseType) {
-            if (null != actual.getOffset()) {
-                assertThat(getFullAssertMessage("Limit offset index assertion error: "), actual.getOffset().getIndex(), is(expected.getOffsetParameterIndex()));
-            }
-            if (null != actual.getRowCount()) {
-                assertThat(getFullAssertMessage("Limit row count index assertion error: "), actual.getRowCount().getIndex(), is(expected.getRowCountParameterIndex()));
-            }
-        } else {
-            if (null != actual.getOffset()) {
-                assertThat(getFullAssertMessage("Limit offset value assertion error: "), actual.getOffset().getValue(), is(expected.getOffset()));
-            }
-            if (null != actual.getRowCount()) {
-                assertThat(getFullAssertMessage("Limit row count value assertion error: "), actual.getRowCount().getValue(), is(expected.getRowCount()));
-            }
-        }
-    }
-    
-    private String getFullAssertMessage(final String assertMessage) {
-        StringBuilder result = new StringBuilder(System.getProperty("line.separator"));
-        result.append("SQL Case ID : ");
-        result.append(sqlCaseId);
-        result.append(System.getProperty("line.separator"));
-        result.append("SQL         : ");
-        if (SQLCaseType.Placeholder == sqlCaseType) {
-            result.append(SQLCasesLoader.getInstance().getSupportedPlaceholderSQL(sqlCaseId));
-            result.append(System.getProperty("line.separator"));
-            result.append("SQL Params  : ");
-            result.append(parserResultSetLoader.getParserResult(sqlCaseId).getParameters());
-            result.append(System.getProperty("line.separator"));
-        } else {
-            result.append(SQLCasesLoader.getInstance().getSupportedLiteralSQL(sqlCaseId, parserResultSetLoader.getParserResult(sqlCaseId).getParameters()));
-        }
-        result.append(System.getProperty("line.separator"));
-        result.append(assertMessage);
-        result.append(System.getProperty("line.separator"));
-        return result.toString();
+        new SQLStatementAssert(new SQLParsingEngine(databaseType, sql, getShardingRule()).parse(), sqlCaseId, sqlCaseType).assertSQLStatement();
     }
 }
